+++ Plugins/ReactiveUI/Source/ReactiveUISlate/Public/RuiStyle.h
@@ -109,6 +109,8 @@ namespace RUI
 		{
 			return Set(FName(TEXT("FillColorAndOpacity")), FRuiValue(V));
 		}
+		FRuiStyleBuilder& Clipping(FName V) { return Set(FName(TEXT("Clipping")), FRuiValue(V)); }
+		FRuiStyleBuilder& ToolTipText(const FText& V) { return Set(FName(TEXT("ToolTipText")), FRuiValue(V)); }
 
 		FRuiStyleBuilder& Set(FName Key, FRuiValue V)
 		{
@@ -144,6 +146,9 @@ namespace RUI
 			return Set(FName(TEXT("slot.valign")), FRuiValue(VAlignName(V)));
 		}
 		FRuiSlotBuilder& Fill(float Coefficient) { return Set(FName(TEXT("slot.fill")), FRuiValue(Coefficient)); }
+		FRuiSlotBuilder& ZOrder(int32 Z) { return Set(FName(TEXT("Slot.ZOrder")), FRuiValue(static_cast<int64>(Z))); }
+		FRuiSlotBuilder& Position(const FVector2D& V) { return Set(FName(TEXT("Slot.Position")), FRuiValue(V)); }
+		FRuiSlotBuilder& Size(const FVector2D& V) { return Set(FName(TEXT("Slot.Size")), FRuiValue(V)); }
 
 		FRuiSlotBuilder& Set(FName Key, FRuiValue V)
 		{